Fix/duplicate topic documentations (#5)
* fix: duplicate topic documentation
* fix: renamed package from io.muenchendigital.digiwf to io.muenchendigital.digiwf.asyncapi.docs
* fix: detect properties
