--- digiwf-asyncapi-docs-core/src/test/java/io/muenchendigital/digiwf/Payload.java
@@ -1,10 +1,0 @@
-package io.muenchendigital.digiwf;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Setter
-@Getter
-public class Payload {
-    private String message;
-}
