+++ digiwf-asyncapi-docs-core/src/test/java/io/muenchendigital/digiwf/asyncapi/docs/Payload.java
@@ -1,0 +1,10 @@
+package io.muenchendigital.digiwf.asyncapi.docs;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Setter
+@Getter
+public class Payload {
+    private String message;
+}
