+++ digiwf-asyncapi-docs-core/src/test/java/io/muenchendigital/digiwf/asyncapi/docs/Producer.java
@@ -1,0 +1,14 @@
+package io.muenchendigital.digiwf.asyncapi.docs;
+
+import io.muenchendigital.digiwf.asyncapi.docs.annotations.DocumentAsyncAPI;
+import lombok.extern.slf4j.Slf4j;
+
+@Slf4j
+public class Producer {
+
+    @DocumentAsyncAPI(payload = Payload.class)
+    public Payload sendMessage() {
+        return null;
+    }
+
+}
