--- digiwf-asyncapi-docs-starter/src/main/java/io/muenchendigital/digiwf/properties/DocsProperties.java
@@ -1,16 +1,0 @@
-package io.muenchendigital.digiwf.properties;
-
-import lombok.Getter;
-import lombok.Setter;
-import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.context.annotation.Configuration;
-
-@Setter
-@Getter
-@Configuration
-@ConfigurationProperties(prefix = "io.muenchendigital.digiwf.docs")
-public class DocsProperties {
-    private String basePackage;
-    private String version;
-    private String title;
-}
