--- digiwf-asyncapi-docs-starter/src/main/java/io/muenchendigital/digiwf/properties/KafkaProperties.java
@@ -1,19 +1,0 @@
-package io.muenchendigital.digiwf.properties;
-
-import lombok.Getter;
-import lombok.Setter;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Configuration;
-
-@Setter
-@Getter
-@Configuration
-public class KafkaProperties {
-
-    @Value("${spring.cloud.stream.default-binder}")
-    private String defaultBinder;
-
-    @Value("${spring.cloud.stream.kafka.binder.brokers}")
-    private String broker;
-
-}
