--- example-digiwf-asyncapi-docs/src/main/java/io/muenchendigital/digiwf/ExampleApplication.java
@@ -1,14 +1,0 @@
-package io.muenchendigital.digiwf;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-
-@SpringBootApplication
-public class ExampleApplication {
-
-    public static void main(final String[] args) {
-        SpringApplication.run(ExampleApplication.class, args);
-    }
-
-}
