+++ example-digiwf-asyncapi-docs/src/main/java/io/muenchendigital/digiwf/asyncapi/docs/ExampleApplication.java
@@ -1,0 +1,14 @@
+package io.muenchendigital.digiwf.asyncapi.docs;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+
+@SpringBootApplication
+public class ExampleApplication {
+
+    public static void main(final String[] args) {
+        SpringApplication.run(ExampleApplication.class, args);
+    }
+
+}
