+++ example-digiwf-asyncapi-docs/src/main/java/io/muenchendigital/digiwf/asyncapi/docs/dto/DeploymentEvent.java
@@ -1,0 +1,14 @@
+package io.muenchendigital.digiwf.asyncapi.docs.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class DeploymentEvent {
+    private String deploymentId;
+    private String versionId;
+    private String target;
+    private String file;
+    private String artifactType;
+}
