+++ example-digiwf-asyncapi-docs/src/main/java/io/muenchendigital/digiwf/asyncapi/docs/producer/ProducerConfig.java
@@ -1,0 +1,28 @@
+package io.muenchendigital.digiwf.asyncapi.docs.producer;
+
+import io.muenchendigital.digiwf.asyncapi.docs.annotations.DocumentAsyncAPI;
+import io.muenchendigital.digiwf.asyncapi.docs.dto.MessageDto;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.messaging.Message;
+import reactor.core.publisher.Flux;
+import reactor.core.publisher.Sinks;
+
+import java.util.function.Supplier;
+
+@Slf4j
+@Configuration
+public class ProducerConfig {
+
+    @Bean
+    public Sinks.Many<Message<MessageDto>> createMessageSink() {
+        return Sinks.many().unicast().onBackpressureBuffer();
+    }
+
+    @DocumentAsyncAPI(payload = MessageDto.class)
+    @Bean
+    public Supplier<Flux<Message<MessageDto>>> sendMessage(final Sinks.Many<Message<MessageDto>> messagePublisher) {
+        return messagePublisher::asFlux;
+    }
+}
