--- example-digiwf-asyncapi-docs/src/main/java/io/muenchendigital/digiwf/dto/DeploymentEvent.java
@@ -1,14 +1,0 @@
-package io.muenchendigital.digiwf.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class DeploymentEvent {
-    private String deploymentId;
-    private String versionId;
-    private String target;
-    private String file;
-    private String artifactType;
-}
