--- example-digiwf-asyncapi-docs/src/main/java/io/muenchendigital/digiwf/dto/MessageDto.java
@@ -1,14 +1,0 @@
-package io.muenchendigital.digiwf.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-public class MessageDto {
-    private String message;
-}
