--- example-digiwf-asyncapi-docs/src/main/java/io/muenchendigital/digiwf/functionrouter/FunctionRouterConfiguration.java
@@ -1,18 +1,0 @@
-package io.muenchendigital.digiwf.functionrouter;
-
-import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
-import org.springframework.cloud.function.context.MessageRoutingCallback;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-/**
- * This is not a working example. It's for demonstration purposes only!
- */
-@Configuration
-public class FunctionRouterConfiguration {
-    @Bean
-    @ConditionalOnMissingBean
-    public MessageRoutingCallback customRouter() {
-        return null;
-    }
-}
