+++ example-digiwf-asyncapi-docs/src/main/resources/application.properties
@@ -11,6 +11,6 @@ spring.cloud.stream.bindings.deployments-in-0.destination=kafka-demo-deployments
 spring.cloud.stream.bindings.deployments-in-0.group=kafka-demo
 spring.cloud.stream.bindings.deployArtifact-out-0.destination=kafka-demo-deploy-artifact
 
-io.muenchendigital.digiwf.docs.basePackage=io.muenchendigital.digiwf
+io.muenchendigital.digiwf.docs.basePackage=io.muenchendigital.digiwf.asyncapi.docs
 io.muenchendigital.digiwf.docs.version=1.0.0
 io.muenchendigital.digiwf.docs.title=kafka-example
